pyproject.toml
[tool.poetry]
name = "HM-ScalableRecs"
version = "0.1.0"
description = "Let's build a production-ready recommendation system using the Amazon rating dataset."
authors = ["Luong Nguyen Dinh <luongnguyendinhgreen@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
torch = {version = "^2.6.0+cu124", source = "pytorch"}
torchvision = {version = "^0.21.0+cu124", source = "pytorch"}
torchaudio = {version = "^2.6.0+cu124", source = "pytorch"}
python = "3.11.9"
datasets = "^2.21.0"
pandas = "^2.2.2"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
fire = "^0.7.0"
load-dotenv = "^0.1.0"
tqdm = "^4.67.1"
acryl-datahub = "^0.14.1.12"
evidently = "^0.4.40"
psycopg2 = "^2.9.10"
lightning = "^2.5.1"
qdrant-client = {extras = ["fastembed"], version = "^1.13.2"}
redis = "^4.5.6"
bentoml = "^1.3.21"
python-multipart = "^0.0.20"
litestar = "^2.14.0"
langgraph = "^0.3.2"
langsmith = "^0.3.11"
langchain-anthropic = "^0.3.8"
langchain-openai = "^0.3.7"
dask = {extras = ["complete"], version = "^2025.2.0"}
papermill = "^2.6.0"
nbqa = "^1.9.1"
dvc = {extras = ["s3"], version = "^3.59.1"}
mlflow = "^2.21.2"
transformers = "^4.51.3"
accelerate = "^1.7.0"
fastparquet = "^2024.11.0"
minio = "^7.2.15"
xgboost = "^3.0.2"
pydantic-settings = "^2.9.1"
pyspark = "^4.0.0"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.2.4"
ipywidgets = "^8.1.3"
seaborn = "^0.13.2"
matplotlib = "^3.9.2"
black = "^24.10.0"
isort = "^5.13.2"
dvc = {extras = ["s3"], version = "^3.58.0"}
tensorboard = "^2.18.0"

# [[tool.poetry.source]]
# name = "pytorch_cpu"
# url = "https://download.pytorch.org/whl/cpu"
# priority = "explicit"

[tool.poetry.group.feature-pipeline.dependencies]
psycopg2 = "^2.9.10"
dbt-core = "^1.10.2"
dbt-postgres = "^1.9.0"
apache-airflow = {extras = ["postgres"], version = "^3.0.2"}
apache-airflow-providers-docker = "^4.4.0"
feast = {version = "<0.49.1", extras = ["postgres"]}

[[tool.poetry.source]]
name     = "pytorch"
url      = "https://download.pytorch.org/whl/cu124"
priority = "explicit"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

# [tool.ruff.lint.isort]
# force_sort_within_sections = true
